refactor: split any array into collections and array
